Make time we wait for TERM to work config option
Some of our apps take considerably longer than 10 seconds to shut down,
so lets make it configurable how long we wait for TERM to work before we
move on to full nuclear KILL.

This is modeled on the setting for launch_timeout just above, so should
behave similarly. The default is still 10 seconds.

Added a note in the docs about missing tests for this new config option.